Rename kiwixlib to libkiwix
