+++ lib/src/main/cpp/library.cpp
@@ -19,7 +19,7 @@
 
 
 #include <jni.h>
-#include "org_kiwix_kiwixlib_Library.h"
+#include "org_kiwix_libkiwix_Library.h"
 
 #include "library.h"
 #include "reader.h"
