+++ lib/src/main/java/org/kiwix/libkiwix/JNIKiwixBool.java
@@ -17,7 +17,7 @@
  * MA 02110-1301, USA.
  */
 
-package org.kiwix.kiwixlib;
+package org.kiwix.libkiwix;
 
 public class JNIKiwixBool
 {
