+++ lib/src/main/java/org/kiwix/libkiwix/JNIKiwixException.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2013 Emmanuel Engelhart <kelson@kiwix.org>
+ * Copyright (C) 2017 Matthieu Gautier <mgautier@kymeria.org>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU  General Public License as published by
@@ -17,9 +17,11 @@
  * MA 02110-1301, USA.
  */
 
-package org.kiwix.kiwixlib;
+package org.kiwix.libkiwix;
 
-public class JNIKiwixInt
+public class JNIKiwixException extends Exception
 {
-  public int value;
+  public JNIKiwixException(String message) {
+    super(message);
+  }
 }
