+++ lib/src/main/java/org/kiwix/libkiwix/JNIKiwixServer.java
@@ -17,10 +17,10 @@
  * MA 02110-1301, USA.
  */
 
-package org.kiwix.kiwixlib;
+package org.kiwix.libkiwix;
 
-import org.kiwix.kiwixlib.JNIKiwixException;
-import org.kiwix.kiwixlib.Library;
+import org.kiwix.libkiwix.JNIKiwixException;
+import org.kiwix.libkiwix.Library;
 
 public class JNIKiwixServer
 {
