+++ lib/src/main/java/org/kiwix/libkiwix/Manager.java
@@ -17,9 +17,9 @@
  * MA 02110-1301, USA.
  */
 
-package org.kiwix.kiwixlib;
+package org.kiwix.libkiwix;
 
-import org.kiwix.kiwixlib.Library;
+import org.kiwix.libkiwix.Library;
 
 public class Manager
 {
